modularizing codebase
